// components-pro-display/form/enum.tsx
export enum LabelAlign {
  left = 'left',
  center = 'center',
  right = 'right',
}

export enum LabelLayout {
  horizontal = 'horizontal',
  vertical = 'vertical',
  placeholder = 'placeholder',
  none = 'none',
  float = 'float',
}

export enum RequiredMarkAlign {
  left = 'left',
  right = 'right',
}